refactor(playground): move route loaders and link builders out of App.tsx
App.tsx should only export components; exporting loaders and paths from
it required react-refresh eslint suppressions. They now live in
src/lib/app-routing.ts, which tests import directly.
